sync doctr to the upstream main

+++ doctr/models/recognition/transformer/tensorflow.py
@@ -91,7 +91,7 @@ def scaled_dot_product_attention(
     scaled_attention_logits = matmul_qk / tf.math.sqrt(dk)
     # add the mask to the scaled tensor.
     if mask is not None:
-        scaled_attention_logits += (mask * -1e9)
+        scaled_attention_logits += (tf.cast(mask, dtype=q.dtype) * -1e9)
     # softmax is normalized on the last axis (seq_len_k) so that the scores
     # add up to 1.
     attention_weights = tf.nn.softmax(scaled_attention_logits, axis=-1)  # (..., seq_len_q, seq_len_k)
@@ -252,7 +252,7 @@ class Decoder(tf.keras.layers.Layer):
 
         x = self.embedding(x, **kwargs)  # (batch_size, target_seq_len, d_model)
         x *= tf.math.sqrt(tf.cast(self.d_model, x.dtype))
-        x += self.pos_encoding[:, :seq_len, :]
+        x += tf.cast(self.pos_encoding[:, :seq_len, :], dtype=x.dtype)
 
         x = self.dropout(x, **kwargs)
